Fix misuse of {} initialization
Narrowing is supposed to be forbidden in {} initialization, but Chromium
doesn't appear to care upstream.

+++ chromium/base/task/sequence_manager/task_queue_selector.h
@@ -288,16 +288,16 @@ class BASE_EXPORT TaskQueueSelector : public WorkQueueSets::Observer {
           0,
 
           // kVeryHighPriority
-          kMaxVeryHighPriorityStarvationScore,
+          int64_t(kMaxVeryHighPriorityStarvationScore),
 
           // kHighPriority
-          kMaxHighPriorityStarvationScore,
+          int64_t(kMaxHighPriorityStarvationScore),
 
           // kNormalPriority
-          kMaxNormalPriorityStarvationScore,
+          int64_t(kMaxNormalPriorityStarvationScore),
 
           // kLowPriority
-          kMaxLowPriorityStarvationScore,
+          int64_t(kMaxLowPriorityStarvationScore),
 
           // kBestEffortPriority (unused)
           std::numeric_limits<int64_t>::max()};
