+++ chromium/content/renderer/pepper/pepper_plugin_instance_impl.cc
@@ -1987,7 +1987,7 @@ void PepperPluginInstanceImpl::PrintPage(int page_number,
     metafile_ = metafile;
   }
 
-  PP_PrintPageNumberRange_Dev page_range = {page_number, page_number};
+  PP_PrintPageNumberRange_Dev page_range = {uint32_t(page_number), uint32_t(page_number)};
   ranges_.push_back(page_range);
 #endif
 }
